reworking compositionLocal setup for metadata
