+++ gradle/libs.versions.toml
@@ -19,7 +19,7 @@ chucker = "4.1.0"
 coil = "2.7.0"
 compose-lint-checks = "1.4.2"
 coroutines = "1.10.2"
-datastore = "1.1.5"
+datastore = "1.1.4"
 junit = "4.13.2"
 kotlin = "2.1.20"
 kotlin-immutable-collections = "0.3.8"
