+++ lib/src/main/java/com/smileidentity/compose/selfie/OrchestratedSelfieCaptureScreen.kt
@@ -24,9 +24,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.smileidentity.R
 import com.smileidentity.compose.components.ImageCaptureConfirmationDialog
-import com.smileidentity.compose.components.LocalMetadata
 import com.smileidentity.compose.components.ProcessingScreen
-import com.smileidentity.models.v2.Metadatum
+import com.smileidentity.compose.metadata.LocalMetadataProvider
+import com.smileidentity.compose.metadata.Metadatum
 import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDCallback
 import com.smileidentity.util.randomJobId
@@ -52,7 +52,7 @@ fun OrchestratedSelfieCaptureScreen(
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
     extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
-    metadata: SnapshotStateList<Metadatum> = LocalMetadata.current,
+    metadata: SnapshotStateList<Metadatum> = LocalMetadataProvider.current,
     viewModel: SelfieViewModel = viewModel(
         factory = viewModelFactory {
             SelfieViewModel(
