+++ lib/src/main/java/com/smileidentity/compose/selfie/SelfieCaptureScreen.kt
@@ -37,10 +37,10 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.smileidentity.R
 import com.smileidentity.compose.components.ForceBrightness
-import com.smileidentity.compose.components.LocalMetadata
+import com.smileidentity.compose.metadata.LocalMetadataProvider
+import com.smileidentity.compose.metadata.Metadatum
 import com.smileidentity.compose.preview.Preview
 import com.smileidentity.compose.preview.SmilePreviews
-import com.smileidentity.models.v2.Metadatum
 import com.smileidentity.util.randomJobId
 import com.smileidentity.util.randomUserId
 import com.smileidentity.viewmodel.MAX_FACE_AREA_THRESHOLD
@@ -67,7 +67,7 @@ fun SelfieCaptureScreen(
     isEnroll: Boolean = true,
     allowAgentMode: Boolean = true,
     skipApiSubmission: Boolean = false,
-    metadata: SnapshotStateList<Metadatum> = LocalMetadata.current,
+    metadata: SnapshotStateList<Metadatum> = LocalMetadataProvider.current,
     viewModel: SelfieViewModel = viewModel(
         factory = viewModelFactory {
             SelfieViewModel(
