+++ lib/src/main/java/com/smileidentity/compose/theme/SmileThemeSurface.kt
@@ -1,4 +1,4 @@
-package com.smileidentity.compose.components
+package com.smileidentity.compose.theme
 
 import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.MaterialTheme
@@ -6,9 +6,9 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Typography
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.toMutableStateList
-import com.smileidentity.models.v2.Metadata
+import com.smileidentity.compose.metadata.LocalMetadataProvider
+import com.smileidentity.compose.metadata.Metadata
 
 @Composable
 internal fun SmileThemeSurface(
@@ -17,7 +17,7 @@ internal fun SmileThemeSurface(
     content: @Composable () -> Unit,
 ) {
     CompositionLocalProvider(
-        LocalMetadata provides remember { Metadata.default().items.toMutableStateList() },
+        LocalMetadataProvider provides Metadata.default().items.toMutableStateList(),
     ) {
         MaterialTheme(colorScheme = colorScheme, typography = typography) {
             Surface(content = content)
