+++ lib/src/main/java/com/smileidentity/models/PrepUpload.kt
@@ -4,7 +4,7 @@ package com.smileidentity.models
 
 import com.smileidentity.BuildConfig
 import com.smileidentity.SmileID
-import com.smileidentity.models.v2.Metadatum
+import com.smileidentity.compose.metadata.Metadatum
 import com.smileidentity.networking.calculateSignature
 import com.smileidentity.util.getCurrentIsoTimestamp
 import com.squareup.moshi.Json
