+++ lib/src/main/java/com/smileidentity/networking/Retrofit.kt
@@ -1,5 +1,7 @@
 package com.smileidentity.networking
 
+import com.smileidentity.compose.metadata.Metadata
+import com.smileidentity.compose.metadata.Metadatum
 import com.smileidentity.models.BiometricKycJobResult
 import com.smileidentity.models.DocumentVerificationJobResult
 import com.smileidentity.models.EnhancedDocumentVerificationJobResult
@@ -9,8 +11,6 @@ import com.smileidentity.models.PartnerParams
 import com.smileidentity.models.SmartSelfieJobResult
 import com.smileidentity.models.SmileIDException
 import com.smileidentity.models.UploadRequest
-import com.smileidentity.models.v2.Metadata
-import com.smileidentity.models.v2.Metadatum
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.JsonQualifier
