+++ lib/src/main/java/com/smileidentity/viewmodel/SelfieViewModel.kt
@@ -15,17 +15,17 @@ import com.smileidentity.R
 import com.smileidentity.SmileID
 import com.smileidentity.SmileIDCrashReporting
 import com.smileidentity.compose.components.ProcessingState
+import com.smileidentity.compose.metadata.LivenessType
+import com.smileidentity.compose.metadata.Metadatum
+import com.smileidentity.compose.metadata.SelfieImageOriginValue.BackCamera
+import com.smileidentity.compose.metadata.SelfieImageOriginValue.FrontCamera
+import com.smileidentity.compose.metadata.asNetworkRequest
 import com.smileidentity.models.AuthenticationRequest
 import com.smileidentity.models.JobType.SmartSelfieAuthentication
 import com.smileidentity.models.JobType.SmartSelfieEnrollment
 import com.smileidentity.models.PartnerParams
 import com.smileidentity.models.PrepUploadRequest
 import com.smileidentity.models.SmileIDException
-import com.smileidentity.models.v2.LivenessType
-import com.smileidentity.models.v2.Metadatum
-import com.smileidentity.models.v2.SelfieImageOriginValue.BackCamera
-import com.smileidentity.models.v2.SelfieImageOriginValue.FrontCamera
-import com.smileidentity.models.v2.asNetworkRequest
 import com.smileidentity.networking.doSmartSelfieAuthentication
 import com.smileidentity.networking.doSmartSelfieEnrollment
 import com.smileidentity.results.SmartSelfieResult
