+++ lib/src/main/java/com/smileidentity/viewmodel/SmartSelfieEnhancedViewModel.kt
@@ -17,15 +17,15 @@ import com.google.mlkit.vision.face.FaceDetectorOptions
 import com.smileidentity.R
 import com.smileidentity.SmileID
 import com.smileidentity.SmileIDCrashReporting
+import com.smileidentity.compose.metadata.LivenessType
+import com.smileidentity.compose.metadata.Metadatum
+import com.smileidentity.compose.metadata.SelfieImageOriginValue.BackCamera
+import com.smileidentity.compose.metadata.SelfieImageOriginValue.FrontCamera
+import com.smileidentity.compose.metadata.asNetworkRequest
+import com.smileidentity.compose.metadata.model
 import com.smileidentity.ml.SelfieQualityModel
 import com.smileidentity.models.v2.FailureReason
-import com.smileidentity.models.v2.LivenessType
-import com.smileidentity.models.v2.Metadatum
-import com.smileidentity.models.v2.SelfieImageOriginValue.BackCamera
-import com.smileidentity.models.v2.SelfieImageOriginValue.FrontCamera
 import com.smileidentity.models.v2.SmartSelfieResponse
-import com.smileidentity.models.v2.asNetworkRequest
-import com.smileidentity.models.v2.model
 import com.smileidentity.networking.doSmartSelfieAuthentication
 import com.smileidentity.networking.doSmartSelfieEnrollment
 import com.smileidentity.results.SmartSelfieResult
